Consolidating fromCharCode for num => letter to lettable operator
