+++ src/app/app.component.html
@@ -1,1 +1,1 @@
-<rx-catch></rx-catch>
+<rx-window></rx-window>
