+++ src/app/visualizations/concat.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { map, take, delay, concat } from 'rxjs/operators';
+import { take, delay, concat } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-concat',
@@ -13,10 +15,7 @@ import { map, take, delay, concat } from 'rxjs/operators';
   `,
 })
 export class RxConcatComponent {
-  first$ = interval(1000).pipe(
-    take(10),
-    map(count => String.fromCharCode(count + 97)),
-  );
+  first$ = interval(1000).pipe(take(10), mapNumberToChar());
   second$ = interval(1000).pipe(delay(500), take(10));
   concatenated$ = this.first$.pipe(concat(this.second$));
 }
