+++ src/app/visualizations/merge-all.ts
@@ -3,6 +3,8 @@ import { Component } from '@angular/core';
 import { interval } from 'rxjs/observable/interval';
 import { mergeAll, take, map } from 'rxjs/operators';
 
+import { mapNumberToChar } from '../mapNumberToChar';
+
 @Component({
   selector: 'rx-merge-all',
   template: `
@@ -15,10 +17,7 @@ import { mergeAll, take, map } from 'rxjs/operators';
 export class RxMergeAllComponent {
   initTime = new Date().getTime();
   lowerOrders = [];
-  higherOrder$ = interval(1000).pipe(
-    take(4),
-    map(val => String.fromCharCode(val + 97)),
-  );
+  higherOrder$ = interval(1000).pipe(take(4), mapNumberToChar());
   firstOrder$ = this.higherOrder$.pipe(
     map(val => {
       const lowerOrder = interval(1000).pipe(
