+++ src/app/visualizations/merge-map.ts
@@ -3,6 +3,8 @@ import { Component } from '@angular/core';
 import { interval } from 'rxjs/observable/interval';
 import { mergeMap, map, take } from 'rxjs/operators';
 
+import { mapNumberToChar } from '../mapNumberToChar';
+
 @Component({
   selector: 'rx-merge-map',
   template: `
@@ -15,10 +17,7 @@ import { mergeMap, map, take } from 'rxjs/operators';
 export class RxMergeMapComponent {
   initTime = new Date().getTime();
   lowerOrders = [];
-  higherOrder$ = interval(1000).pipe(
-    take(4),
-    map(val => String.fromCharCode(val + 97)),
-  );
+  higherOrder$ = interval(1000).pipe(take(4), mapNumberToChar());
   firstOrder$ = this.higherOrder$.pipe(
     mergeMap(val => {
       const lowerOrder = interval(1000).pipe(
