+++ src/app/visualizations/single.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { single, map, take } from 'rxjs/operators';
+import { single, take } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-single',
@@ -12,9 +14,6 @@ import { single, map, take } from 'rxjs/operators';
   `,
 })
 export class RxSingleComponent {
-  input$ = interval(1000).pipe(
-    take(3),
-    map(val => String.fromCharCode(val + 97)),
-  );
+  input$ = interval(1000).pipe(take(3), mapNumberToChar());
   output$ = this.input$.pipe(single(val => val === 'b'));
 }
