+++ src/app/visualizations/start-with.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { startWith, map, take } from 'rxjs/operators';
+import { startWith, take } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-start-with',
@@ -12,9 +14,6 @@ import { startWith, map, take } from 'rxjs/operators';
   `,
 })
 export class RxStartWithComponent {
-  input$ = interval(1000).pipe(
-    take(3),
-    map(val => String.fromCharCode(val + 97)),
-  );
+  input$ = interval(1000).pipe(take(3), mapNumberToChar());
   output$ = this.input$.pipe(startWith('s'));
 }
