+++ src/app/visualizations/take-last.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { takeLast, map, take } from 'rxjs/operators';
+import { takeLast, take } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-take-last',
@@ -12,9 +14,6 @@ import { takeLast, map, take } from 'rxjs/operators';
   `,
 })
 export class RxTakeLastComponent {
-  input$ = interval(1000).pipe(
-    map(val => String.fromCharCode(val + 97)),
-    take(4),
-  );
+  input$ = interval(1000).pipe(mapNumberToChar(), take(4));
   output$ = this.input$.pipe(takeLast(2));
 }
