+++ src/app/visualizations/throttle-time.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { throttleTime, take, map } from 'rxjs/operators';
+import { throttleTime, take } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-throttle-time',
@@ -12,9 +14,6 @@ import { throttleTime, take, map } from 'rxjs/operators';
   `,
 })
 export class RxThrottleTimeComponent {
-  input$ = interval(500).pipe(
-    map(val => String.fromCharCode(val + 97)),
-    take(10),
-  );
+  input$ = interval(500).pipe(mapNumberToChar(), take(10));
   output$ = this.input$.pipe(throttleTime(1500));
 }
