+++ src/app/visualizations/window-when.ts
@@ -3,6 +3,8 @@ import { Component } from '@angular/core';
 import { interval } from 'rxjs/observable/interval';
 import { windowWhen, map, take, mapTo, mergeAll } from 'rxjs/operators';
 
+import { mapNumberToChar } from '../mapNumberToChar';
+
 @Component({
   selector: 'rx-window-when',
   template: `
@@ -14,10 +16,7 @@ import { windowWhen, map, take, mapTo, mergeAll } from 'rxjs/operators';
 })
 export class RxWindowWhenComponent {
   // I don't understand this one fully
-  input$ = interval(1000).pipe(
-    map(val => String.fromCharCode(val + 97)),
-    take(10),
-  );
+  input$ = interval(1000).pipe(mapNumberToChar(), take(10));
   notifier$ = interval(2500).pipe(mapTo('w'));
   output$ = this.input$.pipe(
     windowWhen(() => interval(2500)),
