+++ src/app/visualizations/with-latest-from.ts
@@ -1,7 +1,9 @@
 import { Component } from '@angular/core';
 
 import { interval } from 'rxjs/observable/interval';
-import { withLatestFrom, map, take, skip } from 'rxjs/operators';
+import { withLatestFrom, take, skip } from 'rxjs/operators';
+
+import { mapNumberToChar } from '../mapNumberToChar';
 
 @Component({
   selector: 'rx-with-latest-from',
@@ -13,10 +15,7 @@ import { withLatestFrom, map, take, skip } from 'rxjs/operators';
   `,
 })
 export class RxWithLatestFromComponent {
-  letters$ = interval(1500).pipe(
-    map(val => String.fromCharCode(val + 97)),
-    take(7),
-  );
+  letters$ = interval(1500).pipe(mapNumberToChar(), take(7));
   numbers$ = interval(1000).pipe(skip(1), take(8));
   combined$ = this.letters$.pipe(
     withLatestFrom(this.numbers$, (letter, number) => letter + number),
