+++ src/index.html
@@ -9,6 +9,6 @@
   <link rel="icon" type="image/x-icon" href="favicon.ico">
 </head>
 <body>
-  <app-root>Loading...</app-root>
+  <rx-visualizations-app>Loading...</rx-visualizations-app>
 </body>
 </html>
